+++ NEWS.md
@@ -1,3 +1,17 @@
+CohortMethod 5.2.1
+==================
+
+Changes:
+
+1. Ask to delete files in output folder when calling `runCmAnalyses()` with different analyses settings than those used to create the files. Also cleaning the cache.
+
+Bugfixes:
+
+1. Fixed bug in parsing covariate filter settings for balance.
+
+2. Updated vignettes to use latest `Capr` functions.
+
+
 CohortMethod 5.2.0
 ==================